BSE-2631/2633: Schema evolution tests (#7241)
* Create_all_tables -> create_tables

* Add type hints and capture wrapped function output

* Convert table_path tests

* Only create table if it doesn't exist

* Don't try to delete non-existent schema

* Bugfixes

* Convert test_table_path

* Convert the rest of the bodosql iceberg tests

* Fix bugs in iceberg helpers

* Convert test_iceberg.py

* Only check size if reading columns

* Update numpy dependency and add clang-tools as a dev dependency

* add clang-tools for arm64

* [Run CI]

* Fix test_stream_iceberg_write

* Fix merge_into_parity

* Mark test as one rank

* Fix another test

* Review comment

* Remove duplication

* Add comment

* Return database_schema from this run of create iceberg table

* Create infra and nullable tables

* Add more schema evolution tables

* Add struct field evolution tables

* Review comments

* Review comments

* Add unsupported struct operations table

* Cleanup

* Add marker

* [Run CI]

* Add combo tables

* Add adversarial table

* Cleanup

* Remove print

* Add imports

* Add read test

* Add write test

* Add marks [run ci]

* Handle unsupported specially [run ci]

* Undo delete [run ci]

* Update spark runtime version

* Update spark runtime version

* New test modifications

* Add tablename for schema_evolution_eg_table

* Fix new tests

* Another fix

* Fix another test

* Fix typos and add docstrings

* Refactor create iceberg table with append helper

* Add data after transformations

* Bugfix

* Convert the types we can to pyarrow

* Fix test

* Create spark in iceberg_database_fixture

* Update from develop

* Add filter pushdown and limit pushdown tests

* Use renamed column in filter

* [Run CI]

* Updated log message

* [Run CI]

* Fix test [run ci]

* Fix type annotation

* Add docstring

* Only write once

* Add comment

* Review comments

* Add equality options

* Add barrier to cleanup

* Add adversarial filter pushdown test

* Do column rename and write

* Dict->dict, List->list

* Adjust filter pushdown test columns value

* Add docstrings

* [Run CI]

---------

+++ bodo/tests/test_iceberg.py
@@ -1439,7 +1439,6 @@ def _verify_pq_schema_in_files(
             )
             data_files = list(set(data_files) - data_files_before_write)
             assert all(os.path.isfile(file) for file in data_files)
-
             # Verify that all the parquet files have the correct metadata:
             for data_file in data_files:
                 pq_file = pq.ParquetFile(data_file)
@@ -1538,7 +1537,7 @@ ICEBERG_FIELD_IDS_IN_PQ_SCHEMA_TEST_PARAMS: list[str, pa.Schema] = [
             [
                 pa.field(
                     "A",
-                    pa.struct([pa.field("a", pa.int64()), pa.field("b", pa.int64())]),
+                    pa.struct([pa.field("a", pa.int32()), pa.field("b", pa.int64())]),
                 ),
                 pa.field(
                     "B",
